+++ src/main.js
@@ -1,12 +1,11 @@
 import { getIssueItemTpl, getIssueTpl, getLabelItemTpl, getLabelTpl } from './tpl';
-import { fetchIssues, fetchLabels } from './common/api';
-import { renderElement, selectElement } from './utils/render';
+import { fetchIssues } from './common/api';
 import { OPEN, CLOSED } from './constants/status';
 import { SELECTOR } from './constants/selector';
 import { EVENT } from './constants/event';
-import { bindEvent, setEventListener } from './utils/event';
 import { getIssuesWithStatus } from './utils/status';
-import { createIssue, createIssueList } from './utils/template';
+import { createIssue, createIssueList, updateIssuesTemplate, updateIssueList } from './utils/template';
+import { selectAllElement } from './utils/render';
 
 
 
@@ -19,7 +18,20 @@ const closedIssueCount = closedIssues.length;
 const getIssueListTableTemplate = issues => issues.map(issue => getIssueItemTpl(issue)).join('');
 
 createIssue(getIssueTpl({openIssueCount, closedIssueCount}));
-setEventListener(issues);
+
+const elements = selectAllElement(SELECTOR.STATUS_TAB);
+const OPEN_COUNT_CLASS = '_open_count'
+elements.forEach(element => {
+  element.addEventListener(EVENT.CLICK, e => {
+    const status = e.target.classList.contains(OPEN_COUNT_CLASS) ? OPEN : CLOSED;
+    const clickedIssues = getIssuesWithStatus(issues, status);
+    const openIssueCount = getIssuesWithStatus(issues, OPEN).length;
+    const closedIssueCount = getIssuesWithStatus(issues, CLOSED).length;
+    updateIssuesTemplate(openIssueCount, closedIssueCount, status);
+    updateIssueList(clickedIssues);
+  });
+});
+
 createIssueList(getIssueListTableTemplate(issues));
